Fix SQL Injection vulnerability in helpdesk selector
The `modules/helpdesk/selector.php` script was vulnerable to SQL Injection
via the `project_company`, `comp`, `company_id`, and `task_project`
parameters. These variables were directly concatenated into SQL strings
without sufficient sanitization.

This commit casts these inputs to integers `(int)` before use, ensuring
only safe numeric values are passed to the query builder.

A regression test `tests/test_security_selector.php` has been added to
verify the fix and prevent future regressions.

+++ modules/helpdesk/selector.php
@@ -30,7 +30,7 @@ function selPermWhere( $table, $idfld ) {
 $debug = false;
 $callback = dPgetParam( $_GET, 'callback', 0 );
 $table = dPgetParam( $_GET, 'table', 0 );
-$comp=dPgetParam($_GET, 'comp', 0);
+$comp = (int) dPgetParam($_GET, 'comp', 0);
 
 $ok = $callback & $table;
 
@@ -51,14 +51,14 @@ case 'companies':
 case 'departments':
 // known issue: does not filter out denied companies
 	$title = 'Department';
-	$company_id = dPgetParam( $_GET, 'company_id', 0 );
+	$company_id = (int) dPgetParam( $_GET, 'company_id', 0 );
 	//$ok &= $company_id;  // Is it safe to delete this line ??? [kobudo 13 Feb 2003]
 	//$where = selPermWhere( 'companies', 'company_id' );
 	$where = "dept_company = company_id ";
 	$where .= "\nAND ".selPermWhere( 'departments', 'dept_id' );
 
 	$table .= ", companies, permissions";
-	$hide_company = dPgetParam( $_GET, 'hide_company', 0 );
+	$hide_company = (int) dPgetParam( $_GET, 'hide_company', 0 );
 	if ( $hide_company == 1 ){
 		$select = "dept_id, dept_name";
 	}else{
@@ -77,7 +77,7 @@ case 'forums':
 	$order = 'forum_name';
 	break;
 case 'projects':
-	$project_company = dPgetParam( $_GET, 'project_company', 0 );
+	$project_company = (int) dPgetParam( $_GET, 'project_company', 0 );
 
 	$title = 'Project';
 	$select = 'project_id,project_name';
@@ -87,7 +87,7 @@ case 'projects':
 	$table .= ", permissions";
 	break;
 case 'tasks':
-	$task_project = dPgetParam( $_GET, 'task_project', 0 );
+	$task_project = (int) dPgetParam( $_GET, 'task_project', 0 );
 
 	$title = 'Task';
 	$select = 'task_id,task_name';
